Return back `balances.rs`
